requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
requests==2.32.4
streamlit==1.46.1